+++ jdaviz/configs/mosviz/plugins/viewers.py
@@ -1,9 +1,10 @@
+import numpy as np
+from astropy import units as u
 from glue.core import BaseData
 from glue_jupyter.bqplot.image import BqplotImageView
 from glue_jupyter.bqplot.profile import BqplotProfileView
 from glue_jupyter.table import TableViewer
 from specutils import Spectrum1D
-from astropy import units as u
 
 from jdaviz.core.events import (AddDataToViewerMessage,
                                 RemoveDataFromViewerMessage,
@@ -11,6 +12,7 @@ from jdaviz.core.events import (AddDataToViewerMessage,
 from jdaviz.core.registries import viewer_registry
 from jdaviz.core.freezable_state import FreezableBqplotImageViewerState
 from jdaviz.configs.default.plugins.viewers import JdavizViewerMixin
+from jdaviz.configs.imviz.helper import layer_is_image_data
 
 __all__ = ['MosvizProfileView', 'MosvizImageView', 'MosvizProfile2DView',
            'MosvizTableViewer']
@@ -140,6 +142,75 @@ class MosvizProfile2DView(BqplotImageView, JdavizViewerMixin):
         self._initialize_toolbar_nested()
         # Setup viewer option defaults
         self.state.aspect = 'auto'
+
+        if self.jdaviz_app.config == 'specviz2d':
+            self.label_mouseover = None
+            self.add_event_callback(self.on_mouse_or_key_event,
+                                    events=['mousemove', 'mouseenter', 'mouseleave'])
+
+    def on_mouse_or_key_event(self, data):
+
+        # Find visible layers
+        visible_layers = [layer for layer in self.state.layers
+                          if (layer.visible and layer_is_image_data(layer.layer))]
+
+        if len(visible_layers) == 0:
+            return
+
+        active_layer = visible_layers[-1]
+
+        if self.label_mouseover is None:
+            if 'g-coords-info' in self.session.application._tools:
+                self.label_mouseover = self.session.application._tools['g-coords-info']
+            else:
+                return
+
+        if data['event'] == 'mousemove':
+            # Display the current cursor coordinates (both pixel and world) as
+            # well as data values. For now we use the first dataset in the
+            # viewer for the data values.
+
+            # Extract first dataset from visible layers and use this for coordinates - the choice
+            # of dataset shouldn't matter if the datasets are linked correctly
+            image = active_layer.layer
+            self.label_mouseover.icon = self.jdaviz_app.state.layer_icons.get(active_layer.layer.label)  # noqa
+
+            # TODO: If this viewer is going to support dithered images link by WCS,
+            #       data coordinates need to be handled like Imviz.
+            # Extract data coordinates - these are pixels in the reference image
+            x = data['domain']['x']
+            y = data['domain']['y']
+
+            if x is None or y is None:  # Out of bounds
+                self.label_mouseover.pixel = ""
+                self.label_mouseover.reset_coords_display()
+                self.label_mouseover.value = ""
+                return
+
+            maxsize = int(np.ceil(np.log10(np.max(image.shape)))) + 3
+            fmt = 'x={0:0' + str(maxsize) + '.1f} y={1:0' + str(maxsize) + '.1f}'
+            self.label_mouseover.pixel = (fmt.format(x, y))
+
+            # TODO: How do display sky coordinates for 2D spectrum?
+            self.label_mouseover.reset_coords_display()
+
+            # Extract data values at this position.
+            # TODO: for now we just use the first visible layer but we should think
+            # of how to display values when multiple datasets are present.
+            if (x > -0.5 and y > -0.5
+                    and x < image.shape[1] - 0.5 and y < image.shape[0] - 0.5
+                    and hasattr(active_layer, 'attribute')):
+                attribute = active_layer.attribute
+                value = image.get_data(attribute)[int(round(y)), int(round(x))]
+                unit = image.get_component(attribute).units
+                self.label_mouseover.value = f'{value:+10.5e} {unit}'
+            else:
+                self.label_mouseover.value = ''
+
+        elif data['event'] == 'mouseleave' or data['event'] == 'mouseenter':
+            self.label_mouseover.pixel = ""
+            self.label_mouseover.reset_coords_display()
+            self.label_mouseover.value = ""
 
     def data(self, cls=None):
         return [layer_state.layer.get_object(cls=cls or self.default_class)